Conflicts:
	1. Suit 'em up Game Development/SuitEmUp/Debug/SuitEmUp.exe
	1. Suit 'em up Game Development/SuitEmUp/Debug/SuitEmUp.ilk
	1. Suit 'em up Game Development/SuitEmUp/Debug/SuitEmUp.pdb
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp.sdf
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp.v11.suo
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/CL.read.1.tlog
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/CL.write.1.tlog
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Collider.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/CollisionManager.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Config.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Customize.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/EnemyObject.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/EnemyProjectile.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Engine.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Game.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/GameObject.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/GameObjectManager.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/InputManager.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Level.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/MainMenu.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/MovingGameObject.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Options.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/PlayerObject.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/PlayerProjectile.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Projectile.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Spawner.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Sprite.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/SpriteManager.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/StateManager.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/SuitEmUp.Build.CppClean.log
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/SuitEmUp.log
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/TitleScreen.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Truck.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/Utils.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/cl.command.1.tlog
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/link.command.1.tlog
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/link.read.1.tlog
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/link.write.1.tlog
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/main.obj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/vc110.idb
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Debug/vc110.pdb
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/SuitEmUp.vcxproj
	1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/SuitEmUp.vcxproj.filters

+++ 1. Suit 'em up Game Development/SuitEmUp/SuitEmUp/Button.cpp
@@ -37,8 +37,8 @@ std::string Button::Update()
 
 	if(m_Shape == "Circle")
 	{
-		int dx = (m_input->GetMousePos().x - m_position.x);
-		int dy = (m_input->GetMousePos().y - m_position.y);
+		float dx = (m_input->GetMousePos().x - m_position.x);
+		float dy = (m_input->GetMousePos().y - m_position.y);
 
 		float hyp = sqrt((dx*dx)+(dy*dy));
